refactor: reorganize GitHub PR diff fetching logic by moving functionality to a new module and updating imports
